formato de columnas, nuevo componente btn
