+++ src/screens/TaskDetails.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import {View, Text, StyleSheet} from 'react-native';
 import GoBackBtn from '../components/GoBackBtn';
 import {Task as TTask} from '../types/types';
+import MyButton from '../components/MyButton';
 
 const TaskDetails = props => {
   const {title, description, category, status} = props.route.params;
@@ -15,6 +16,9 @@ const TaskDetails = props => {
         <Text style={styles.text}>{description}</Text>
         <Text style={styles.text}>{category}</Text>
         <Text style={styles.text}>{status}</Text>
+        <View style={styles.options}>
+            <MyButton title="Delete task" onPress={() => {}} />
+        </View>
       </View>
     </View>
   );
@@ -28,5 +32,8 @@ const styles = StyleSheet.create({
   },
   text: {
     marginBottom: 10,
+  },
+  options: {
+    alignItems: 'center',
   },
 });
