+++ src/screens/TaskList.tsx
@@ -5,14 +5,14 @@ import Task from '../components/Task';
 
 const tasks: Tasks = [
   {
-    title: 'Lorem ipsum dolor sit amet',
+    title: 'Lorem ipsum dolor sit amet 1',
     description:
       'Lorem ipsum dolor sit amet consectetur adipisicing elit. Explicabo aperiam minus placeat ea doloribus sequi. Explicabo debitis, corrupti consectetur repudiandae rerum labore dolorem, tenetur odio dolor neque accusantium molestias distinctio?',
     category: 'doloribus',
     status: false,
   },
   {
-    title: 'Lorem ipsum dolor sit amet',
+    title: 'Lorem ipsum dolor sit amet 2',
     description:
       'Lorem ipsum dolor sit amet consectetur adipisicing elit. Explicabo aperiam minus placeat ea doloribus sequi. Explicabo debitis, corrupti consectetur repudiandae rerum labore dolorem, tenetur odio dolor neque accusantium molestias distinctio?',
     category: 'doloribus',
